refactor(challenge): split filter tests and cover empty-result case per #1019 review
Address review feedback from @kenan-it-academy: move language filter tests into dedicated test classes (InMemoryChallengeRepositoryFilterTest and ChallengeControllerFilterTest) for single-responsibility, and add an explicit test for the 'valid filter, no matching challenges' case (200 with empty list, per REST conventions for collection endpoints).

Refs #1010
